good
All data captured but the check for next-page css does not fail and we wait for it to load and timeout.

What should happen is the locator should return false and we bail, that's what I thought would happen.

Test timeout of 120000ms exceeded.
Error: page.click: Target closed
=========================== logs ===========================
waiting for locator('.NextLast a:first-child')
============================================================

   at ../pages/search.ts:67

  65 |     const css = '.NextLast a:first-child'
  66 |     if (this.page.locator(css)) {
> 67 |       await this.page.click(css)
     |                       ^
  68 |       await this.page.waitForLoadState()
  69 |       await this.capture()
  70 |       await this.page.screenshot({path:`./test-results/screenshot-page-${this.pindex()}.png`})